backend/requirements.txt
alembic==1.16.2
APScheduler==3.10.4
blinker==1.9.0
CacheControl==0.14.3
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==46.0.2
dnspython==2.8.0
email-validator==2.3.0
firebase-admin==6.5.0
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gevent==25.9.1
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
msgpack==1.1.1
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.8.0
pyparsing==3.2.5
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.0.1
WTForms==3.2.1
zope.event==6.0
zope.interface==8.0.1
